Simulation:  fixed some numerical issues with the encoders
Still need to check what the correct values for _cml, _cmr and _b should
be.  0.158 for cml/cmr seems to work.  0.158 seems to result in a match
between robot's calculated theta and theta provided by the simulator,
but need to check _b

+++ robot/sim.h
@@ -9,8 +9,8 @@ typedef struct
 	s16 m1;
 	s16 m2;
 	u16 servo[8];
-	s16 enc_ab;
-	s16 enc_cd;
+	double enc_ab;
+	double enc_cd;
 	SVPStatus svp_status;
 	u32 elapsed_milliseconds;
 	u08* rx_ring_buffer;
